requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
pattern==2.6
django-dotenv==1.3.0
python-amazon-product-api==0.2.8